agents/pyproject.toml

[project]
name = "agents"
version = "0.1.0"
description = "Supply Guardian Agents"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "google-adk>=0.1.0",
    "google-generativeai",
    "httpx",
    "nest-asyncio",
    "python-dotenv",
    "uvicorn",
]

[tool.uv]
package = true
